+++ CHANGELOG.md
@@ -53,6 +53,7 @@ Change categories:
 - utils: Added `pldm_edac_crc32_extend()`
 - base: Added `decode_pldm_base_negotiate_transfer_params_req()`
 - base: Added `encode_pldm_base_negotiate_transfer_params_resp()`
+- platform: Added encode req & decode resp APIs for SetNumericSensorEnable
 
 ### Changed